chore: extract not common parts from SettingsVariables to designated places
