+++ src/components/pages/Tests/TestDetails/TestSettings/SettingsVariables/Arguments.styled.tsx
