+++ src/components/pages/Tests/TestDetails/TestSettings/SettingsVariables/Arguments.tsx
@@ -10,6 +10,13 @@ import {Test} from '@models/test';
 
 import {ConfigurationCard, CopyCommand, notificationCall} from '@molecules';
 
+import {
+  dash,
+  doubleDash,
+  space,
+  stringSpace,
+} from '@organisms/EntityDetails/EntityDetailsContent/Settings/SettingsVariables/utils';
+
 import {Permissions, usePermission} from '@permissions/base';
 
 import {useUpdateTestMutation} from '@services/tests';
@@ -22,7 +29,6 @@ import {externalLinks} from '@utils/externalLinks';
 import {displayDefaultNotificationFlow} from '@utils/notification';
 
 import {ArgumentsWrapper} from './Arguments.styled';
-import {dash, doubleDash, space, stringSpace} from './utils';
 
 type ArgumentsFormValues = {
   args: string;
