+++ src/components/pages/Tests/TestDetails/TestSettings/SettingsVariables/index.ts
@@ -1,0 +1,1 @@
+export {default} from './SettingsVariables';
